Feat: connection with the database
