+++ index.php
@@ -47,6 +47,7 @@
 
 <?php
 
+require_once 'libs/database.php';
 require_once 'libs/router.php';
 
 $router = new Router();
